add test case

+++ backend/src/main/java/demo/application/backend/weather/repository/LocationRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface LocationRepository {
 
 	List<Location> searchLocation(String searchClause);
-	Location findLocation(float lat, float lon);
+	Location findLocation(double lat, double lon);
 }
